- split out the main movement calculation out of P_PlayerInputBoat and used the original code for providing a replacement.

+++ source/games/duke/src/hudweapon_r.cpp
@@ -92,15 +92,15 @@ void displaymasks_r(int snum)
 //
 //---------------------------------------------------------------------------
 
-void ShowMotorcycle(double x, double y, int tilenum, int shade, int orientation, int p, int a)
+void ShowMotorcycle(double x, double y, int tilenum, int shade, int orientation, int p, double a)
 {
-	hud_drawsprite(x, y, 34816, a, tilenum, shade, p, 2 | orientation);
+	hud_drawsprite(x, y, 34816, int(a), tilenum, shade, p, 2 | orientation);
 }
 
 
-void ShowBoat(double x, double y, int tilenum, int shade, int orientation, int p, int a)
+void ShowBoat(double x, double y, int tilenum, int shade, int orientation, int p, double a)
 {
-	hud_drawsprite(x, y, 66048, a, tilenum, shade, p, 2 | orientation);
+	hud_drawsprite(x, y, 66048, int(a), tilenum, shade, p, 2 | orientation);
 }
